Revamp default theme
This removes Bootstrap, jQuery, and normalize.css in favor of custom
styling that uses a bit of Flexbox and CSS Grid. All templates tweaked
accordingly.

Overall, this replaces thousands of lines of CSS (including ~100 lines
of custom CSS) with < 300 lines of custom CSS. The site is now
JavaScript-free as well.